split queue components into folders
