+++ packages/app/components/Queue/Student/QuestionForm.tsx
@@ -3,8 +3,8 @@ import { Modal, Alert, Button, Input, Radio } from "antd";
 import { RadioChangeEvent } from "antd/lib/radio";
 import { default as React, ReactElement, useEffect, useState } from "react";
 import styled from "styled-components";
-import { useLocalStorage } from "../../hooks/useLocalStorage";
-import { toOrdinal } from "../../utils/ordinal";
+import { useLocalStorage } from "../../../hooks/useLocalStorage";
+import { toOrdinal } from "../../../utils/ordinal";
 
 const Container = styled.div`
   max-width: 960px;
