+++ packages/app/components/Queue/Student/StudentInfoCard.tsx
