+++ packages/app/components/Queue/Student/StudentQueueCard.tsx
@@ -2,8 +2,8 @@ import { Question } from "@template/common";
 import { Card, Col } from "antd";
 import { ReactElement } from "react";
 import styled from "styled-components";
-import { getWaitTime } from "../../utils/TimeUtil";
-import { CenterRow, Text } from "./QueueCardSharedComponents";
+import { getWaitTime } from "../../../utils/TimeUtil";
+import { CenterRow, Text } from "../QueueCardSharedComponents";
 
 const HorizontalStudentCard = styled(Card)`
   margin-bottom: 8px;
