+++ packages/app/components/Queue/Student/StudentRemovedFromQueueModal.tsx
