+++ packages/app/components/Queue/TA/TABanner.tsx
@@ -9,7 +9,7 @@ import {
 import { Avatar, Col, Popconfirm, Row } from "antd";
 import { ReactElement } from "react";
 import styled from "styled-components";
-import Banner, { BannerButton, BannerDangerButton } from "./Banner";
+import Banner, { BannerButton, BannerDangerButton } from "../Banner";
 
 const Bold = styled.span`
   font-weight: bold;
