+++ packages/app/components/Queue/TA/TAQueueList.tsx
@@ -8,16 +8,16 @@ import {
 import { Card, Col, Row, Space, Tooltip } from "antd";
 import { ReactElement, useCallback, useState } from "react";
 import styled from "styled-components";
-import { useProfile } from "../../hooks/useProfile";
-import { useQuestions } from "../../hooks/useQuestions";
-import { useQueue } from "../../hooks/useQueue";
+import { useProfile } from "../../../hooks/useProfile";
+import { useQuestions } from "../../../hooks/useQuestions";
+import { useQueue } from "../../../hooks/useQueue";
 import { EditQueueModal } from "./EditQueueModal";
 import {
   QueueInfoColumn,
   QueueInfoColumnButton,
   QueuePageContainer,
   VerticalDivider,
-} from "./QueueListSharedComponents";
+} from "../QueueListSharedComponents";
 import StudentPopupCard from "./StudentPopupCard";
 import TABanner from "./TABanner";
 import TAQueueCard from "./TAQueueCard";
