+++ packages/app/pages/course/[cid]/queue/[qid].tsx
@@ -4,8 +4,8 @@ import { useRouter } from "next/router";
 import { ReactElement } from "react";
 import styled from "styled-components";
 import NavBar from "../../../../components/Nav/NavBar";
-import StudentQueueList from "../../../../components/Queue/StudentQueueList";
-import TAQueueList from "../../../../components/Queue/TAQueueList";
+import StudentQueueList from "../../../../components/Queue/Student/StudentQueueList";
+import TAQueueList from "../../../../components/Queue/TA/TAQueueList";
 import { useQueue } from "../../../../hooks/useQueue";
 import { useRoleInCourse } from "../../../../hooks/useRoleInCourse";
